refactor: move conditional approval form in project.tsx into separate component
